signinscreen auth test all pass - next test components
